backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.4
